predictor/requirements.txt
kafka-python
numpy
pandas